Win: fixing Hunspell compilation; fixing MF boot (Kanban); adding temp file gear function for Win #1386

+++ lib/src/mind/knowledge_graph.cpp
@@ -110,7 +110,9 @@ KnowledgeGraphNode* KnowledgeGraph::getNode(Note* n)
 {
     KnowledgeGraphNode* k = new KnowledgeGraphNode{KnowledgeGraphNodeType::NOTE, n->getName(), notesColor};
     k->setThing(n);
-    k->setCardinality(n->getOutline()->getDirectNoteChildrenCount(n));
+    k->setCardinality(
+        static_cast<unsigned int>(n->getOutline()->getDirectNoteChildrenCount(n))
+    );
 
     return k;
 }
@@ -168,7 +170,9 @@ void KnowledgeGraph::getRelatedNodes(KnowledgeGraphNode* centralNode, KnowledgeS
                 // TODO: reuse and delete - map<Thing*,Node*>
                 k = new KnowledgeGraphNode{KnowledgeGraphNodeType::NOTE, n->getName(), notesColor};
                 k->setThing(n);
-                k->setCardinality(n->getOutline()->getDirectNoteChildrenCount(n));
+                k->setCardinality(
+                    static_cast<unsigned int>(n->getOutline()->getDirectNoteChildrenCount(n))
+                );
                 subgraph.addChild(k);
             }
         }
@@ -220,7 +224,9 @@ void KnowledgeGraph::getRelatedNodes(KnowledgeGraphNode* centralNode, KnowledgeS
             // TODO: reuse and delete - map<Thing*,Node*>
             k = new KnowledgeGraphNode{KnowledgeGraphNodeType::NOTE, n->getName(), notesColor};
             k->setThing(n);
-            k->setCardinality(n->getOutline()->getDirectNoteChildrenCount(n));
+            k->setCardinality(
+                static_cast<unsigned int>(n->getOutline()->getDirectNoteChildrenCount(n))
+            );
             subgraph.addChild(k);
         }
